+++ site_electron/src/pages/CharacterSelect.tsx
@@ -1,0 +1,6 @@
+import { Settings } from "../App";
+import { GrimData } from "./NewGrim";
+
+export default function CharacterSelect({settings, grimData}: {settings: Settings, grimData: GrimData}) {
+    return <h1 style={{backgroundColor: settings.secondaryColour}}>Select Characters layout: {grimData.layout}</h1>
+}