+++ site_electron/changes.md
@@ -1,6 +1,17 @@
 ## New Features
-- Added a nav bar to grim and new grim pages
+- New Grim Page
+    - Script Selection
+    - Player Count Selection
+    - Change Layout Settings (has no effect)
+- Made Settings Preview a Component
+- Added a secondary colour preview to settings
+- Added text colour preview to settings
+- Added Change Name Menu Option (non-functional)
+- Removed Old Warning Comment from Home Page
+- Made player tokens' positions and reminders persist
+- Made character text colour match settings
+- Made available reminders change depending selected script (not custom)
 
-## Bugfixes
-- Fixed papayrus font not working
-- Fixed name tags being misplaced
+## Known Bugs
+- Text preview misaligned
+- Links on Credits page no longer styled